third_party/blink/web_tests/platform/generic/fast/events/wheel/wheelevent-constructor-expected.txt

Tests the constructor of the standard wheel event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testEvent = new WheelEvent('wheel') did not throw exception.
PASS testEvent.__proto__ is WheelEvent.prototype
PASS testEvent.deltaX is 0
PASS testEvent.deltaY is 0
PASS testEvent.deltaZ is 0
PASS testEvent.deltaMode is WheelEvent.DOM_DELTA_PIXEL
PASS testEvent = new WheelEvent('wheel', { deltaX: 10.2, deltaY: 20.3, deltaZ: 30.4, deltaMode: WheelEvent.DOM_DELTA_PAGE }) did not throw exception.
PASS testEvent.__proto__ is WheelEvent.prototype
PASS testEvent.deltaX is 10.2
PASS testEvent.deltaY is 20.3
PASS testEvent.deltaZ is 30.4
PASS testEvent.deltaMode is WheelEvent.DOM_DELTA_PAGE
PASS successfullyParsed is true

TEST COMPLETE
